Split AgentDispatcher into per-domain mixins (Phase 4c: god-file decomposition)
Problem: backend/agents.py was 3837 lines, dominated by the AgentDispatcher
class (~3000 lines, 55 methods) - every LLM dispatch surface in the app
(chat/conversation/image replies, web research, artifact, gitlink, code
sandbox, the Builder loop, the Harness agent, chart/note generation, branch
comparison/synthesis, plus cancellation/approval plumbing) in one class
with no internal boundaries.

Change: split the class into 8 per-domain mixins under a new
backend/agent_dispatch/ package (core, builder, harness, chat, research,
gitlink, code_sandbox, content), composed back into one AgentDispatcher in
backend/agents.py - the third use of the codebase's established
mixin-composition pattern (SceneDocument, SettingsManager). agents.py keeps
its module docstring, every module-level helper (_call_* agents, gitlink
helpers), all constants, bootstrap_provider_state/apply_provider_mode,
register_agents, and the composed class; it shrinks 3837 -> 847 lines.

The engineering constraint this split was built around: the test suite
patches names THROUGH backend.agents' module namespace
(monkeypatch.setattr(agents_module, "_call_chat_agent_stream", ...) 19x,
"VirtualEnvSandbox" 13x, "apply_change_set" 12x, plus 7 more names), and
Python resolves a function's globals in its DEFINING module - so a moved
method that bound those names at import time would silently stop seeing
every such patch. Every moved method therefore accesses agents.py-namespace
names late-bound, via a deferred `from backend import agents as
agents_module` + attribute access at call time - the patch seam survives
with ZERO test changes to any monkeypatch site. Verified mechanically: an
AST comparison of all 55 relocated method bodies against the originals
shows 31 byte-identical and 24 differing EXCLUSIVELY by that late-binding
rewrite (proven by normalizing the rewrite out and re-comparing - zero
logic changes). respond_json/apply_change_set/validate_pending_changes (the
3 remaining patched names) are only called from module-level helpers that
stayed in agents.py, so their seams were never at risk.

Three source-scanning meta-tests hardcoded backend/agents.py as the file
holding AgentDispatcher's methods and needed their scan targets extended
to the new package (the invariants themselves are unchanged and still
enforced, now across the whole mixin set):
- backend/tests/test_dispatch_claim_ordering.py (the no-await-between-
  busy-check-and-claim gate) now scans agents.py + agent_dispatch/*.py.
- backend/tests/test_nonblocking_dispatch.py's two static claim() scans
  glob agent_dispatch/*.py so a future mixin joins automatically.
- tests/test_node_state_migration.py's fallback_ref allowlist entry moved
  from agents.py to core.py, where _dispatch now lives.

No packaging change needed: backend* is already included in pyproject.toml's
setuptools packages.find (verified via a direct find_packages() probe).

Test plan:
- Full backend suite from the repo root (python -m pytest -q): 3073 passed,
  19 skipped, 1 failed - the one failure is the known pre-existing,
  environment-sensitive 2.0s timing assertion in test_providers.py,
  unrelated to this change (reproduces on a clean tree).
- ruff check . and mypy: both clean.
- AST completeness probe: 55/55 methods present across the mixins, zero
  missing, zero extra, zero cross-mixin name collisions.
- Zero module-top backend.agents imports in any mixin (grep-verified) -
  only deferred in-body imports, preserving both patch seams and
  circular-import safety.
