+++ backend/agent_dispatch/__init__.py
@@ -1,0 +1,1 @@
+"""The agent-dispatch package: AgentDispatcher's per-domain mixins."""
