+++ tests/test_node_state_migration.py
@@ -226,10 +226,12 @@ _KNOWN_NON_NODE_FIELD_ACCESS_SHAPES = {
         {"root": "descriptor", "file": "intents_settings_api_provider.py"},
         {"root": "descriptor", "file": "test_model_routing.py"},
         # ADR-018 stage 18.5: _thread_on_fallback's own `fallback_ref` param
-        # (backend/agents.py) is a graphlink_model_catalog.ModelRef -
+        # (backend/agent_dispatch/core.py, inside AgentDispatcher._dispatch -
+        # relocated there from backend/agents.py by the god-file
+        # decomposition) is a graphlink_model_catalog.ModelRef -
         # api_provider's fallback-chain wrapper hands it the model it just
         # substituted in, never a SceneNode.
-        {"root": "fallback_ref", "file": "agents.py"},
+        {"root": "fallback_ref", "file": "core.py"},
         # ADR-018 stage 18.5 review-fix regression test: `catalog[0].provider`
         # is a unified_catalog() result (list[ModelDescriptor]), never a
         # SceneNode collection - test_model_routing.py is entirely about
